Put test of data simualtion in test scenario
